+++ src/components/SideBar/DesktopSideBar.tsx
@@ -2,8 +2,8 @@ import { FoldIcon, LogoIcon } from '@assets';
 import Popup from '@components/Popup';
 import TodoCreateModal from '@components/TodoModal/TodoCreateModal';
 import { useNavigate } from 'react-router-dom';
+import useModalControl from '../../hooks/useModalControl';
 import DesktopSideBarContents from './DesktopSideBarContents';
-import useModalControl from './useModalControl';
 
 interface SideBarProps {
   isOpen: boolean;
