+++ src/components/SideBar/MobileSideBar.tsx
@@ -1,8 +1,8 @@
 import { FoldIcon, HamburgerIcon } from '@assets';
 import Popup from '@components/Popup';
 import TodoCreateModal from '@components/TodoModal/TodoCreateModal';
+import useModalControl from '../../hooks/useModalControl';
 import MobileSideBarContents from './MobileSideBarContents';
-import useModalControl from './useModalControl';
 
 interface MobileSideBarProps {
   isOpen: boolean;
